+++ src/model/problems_resource.cpp
@@ -45,10 +45,10 @@ DECLARE_EXPORT void Resource::updateProblems()
   Date shortageProblemStart;
   bool excessProblem = false;
   bool shortageProblem = false;
-  float curMax(0.0);
-  float shortageQty(0.0);
-  float curMin(0.0);
-  float excessQty(0.0);
+  double curMax(0.0);
+  double shortageQty(0.0);
+  double curMin(0.0);
+  double excessQty(0.0);
   for (loadplanlist::const_iterator iter = loadplans.begin(); 
     iter != loadplans.end(); )
   {
@@ -63,7 +63,7 @@ DECLARE_EXPORT void Resource::updateProblems()
     if (iter!=loadplans.end() && iter->getDate()==f->getDate()) continue;
 
     // Check against minimum target
-    float delta = static_cast<float>(f->getOnhand() - curMin);
+    double delta = f->getOnhand() - curMin;
     if (delta < -ROUNDING_ERROR)
     {
       if (!shortageProblem)
@@ -92,7 +92,7 @@ DECLARE_EXPORT void Resource::updateProblems()
     // the same moment in time.
 
     // Check against maximum target
-    delta = static_cast<float>(f->getOnhand() - curMax);
+    delta = f->getOnhand() - curMax;
     if (delta > ROUNDING_ERROR)
     {
       if (!excessProblem)
